+++ packages/terra-site/src/TestLinks.jsx
@@ -12,6 +12,7 @@ const TestLinks = () => (
       <li><Link to="/tests/base-tests">Base Tests</Link></li>
       <li><Link to="/tests/button-group-tests">Button Group Tests</Link></li>
       <li><Link to="/tests/button-tests">Button Tests</Link></li>
+      <li><Link to="/tests/card-tests">Card Tests</Link></li>
       <li><Link to="/tests/content-container-tests">ContentContainer Tests</Link></li>
       <li><Link to="/tests/date-picker-tests">DatePicker Tests</Link></li>
       <li><Link to="/tests/demographics-banner-tests">DemographicsBanner Tests</Link></li>
@@ -42,6 +43,7 @@ const TestLinks = () => (
       <li><Link to="/tests/status-tests">Status Tests</Link></li>
       <li><Link to="/tests/table-tests">Table Tests</Link></li>
       <li><Link to="/tests/text-tests">Text Tests</Link></li>
+      <li><Link to="/tests/theme-provider-tests">Theme Provider Tests</Link></li>
       <li><Link to="/tests/time-input-tests">Time Input Tests</Link></li>
       <li><Link to="/tests/toggle-tests">Toggle Tests</Link></li>
       <li><Link to="/tests/toggle-button-tests">Toggle Button Tests</Link></li>