+++ packages/terra-i18n/CHANGELOG.md
@@ -4,7 +4,10 @@ Changelog
 Unreleased
 ----------
 ### Changed
+
 * Update nightwatch tests and test scripts.
+
+* Improved terra-i18n example in README.md
 
 1.3.0 - (August 8, 2017)
 ------------------